fix: rename components without breaking feature
