+++ src/components/AccountSelect/AccountSelect.style.tsx
@@ -1,27 +1,25 @@
 import styled from 'styled-components';
 
-import { Flex } from '@/component-library/Flex';
-import { Span } from '@/component-library/Text';
-import { theme } from '@/component-library/theme';
+import { Flex, Span, theme } from '@/component-library';
 
 type StyledListItemSelectedLabelProps = {
   $isSelected: boolean;
 };
 
-const StyledAccountLabelAddress = styled(Span)`
+const StyledAccountItemAddress = styled(Span)`
   text-overflow: ellipsis;
   overflow: hidden;
   white-space: nowrap;
 `;
 
-const StyledAccountLabelName = styled(StyledAccountLabelAddress)<StyledListItemSelectedLabelProps>`
+const StyledAccountItemName = styled(StyledAccountItemAddress)<StyledListItemSelectedLabelProps>`
   color: ${({ $isSelected }) =>
     $isSelected ? theme.tokenInput.list.item.selected.text : theme.tokenInput.list.item.default.text};
 `;
 
-const StyledAccountLabelWrapper = styled(Flex)`
+const StyledAccountItemWrapper = styled(Flex)`
   flex-grow: 1;
   overflow: hidden;
 `;
 
-export { StyledAccountLabelAddress, StyledAccountLabelName, StyledAccountLabelWrapper };
+export { StyledAccountItemAddress, StyledAccountItemName, StyledAccountItemWrapper };
