+++ src/components/AccountSelect/AccountSelect.tsx
@@ -2,7 +2,7 @@ import { InjectedAccountWithMeta } from '@polkadot/extension-inject/types';
 
 import { Item, Select, SelectProps } from '@/component-library';
 
-import { AccountLabel } from './AccountLabel';
+import { AccountItem } from './AccountItem';
 
 type AccountSelectProps = Omit<SelectProps<'modal', InjectedAccountWithMeta>, 'children' | 'type'>;
 
@@ -11,7 +11,7 @@ const AccountSelect = (props: AccountSelectProps): JSX.Element => {
     <Select<'modal', any> {...props} type='modal' modalTitle='select account' size='large'>
       {(data: InjectedAccountWithMeta) => (
         <Item key={data.address} textValue={data.address}>
-          <AccountLabel address={data.address} name={data.meta.name} />
+          <AccountItem address={data.address} name={data.meta.name} />
         </Item>
       )}
     </Select>
